Refactor discord command handling Will be better suited for implementing slashcommands and easy addition of new commands
